src/requirements.txt
Django==1.10
django-debug-toolbar==1.9.1
MySQL-python==1.2.5
olefile==0.45.1
Pillow==4.3.0
selenium==3.10.0
sqlparse==0.2.4
